Appdata sanity checks

+++ lib/handle.c
@@ -150,7 +150,7 @@ static int _pndman_handle_download(pndman_handle *handle)
 
    /* check appdata */
    appdata = _pndman_device_get_appdata(handle->device);
-   if (!strlen(appdata)) {
+   if (!appdata || !strlen(appdata)) {
       curl_free_request(&handle->request);
       return RETURN_FAIL;
    }
@@ -244,7 +244,7 @@ static int _pndman_handle_install(pndman_handle *handle, pndman_repository *loca
 
    /* check appdata */
    appdata = _pndman_device_get_appdata(handle->device);
-   if (!strlen(appdata)) return RETURN_FAIL;
+   if (!appdata || !strlen(appdata)) return RETURN_FAIL;
 
    /* complete handle's path */
    snprintf(tmp, PATH_MAX-1, "%s/%p.tmp", appdata, handle);
@@ -412,7 +412,7 @@ int pndman_handle_free(pndman_handle *handle)
 
       /* check appdata */
       appdata = _pndman_device_get_appdata(handle->device);
-      if (appdata) {
+      if (appdata && strlen(appdata)) {
          snprintf(tmp_path, PATH_MAX-1, "%s/%p.tmp", appdata, handle);
          unlink(tmp_path);
       }
